exclude extension from package if disabled

+++ tools/package.sh
@@ -16,13 +16,18 @@ if [ "$DEBUG_VERSION" = "" ];then
     DEBUG_VERSION="OFF"
 fi
 
+EXTENSIONS=$4
+if [ "$EXTENSIONS" = "" ];then
+    EXTENSIONS="OFF"
+fi
+
 if [ "$SHASUM" = "" ];then
     export SHASUM="shasum"
 fi
 
 
 if [ "$TYPE" = "" ] || [ "$VERSION" = "" ];then
-    echo "Usage: $0 <type> <version> <debug=OFF>"
+    echo "Usage: $0 <type> <version> <debug=OFF> <enable_extensions=OFF>"
     echo "  example: $0 darwin 1.2.3 OFF OFF"
     exit;
 fi
@@ -35,13 +40,15 @@ if [ -d "$PACK" ];then
     exit;
 fi
 
-if [ -d "$PACK_EXT" ];then
+if [ "$EXTENSIONS" = "ON" ];then
+    if [ -d "$PACK_EXT" ];then
     echo "Error: $PACK_EXT already exist. Remove $PACK_EXT first"
     exit;
+    fi
 fi
 
 
-cmake -DCMAKE_BUILD_TYPE=Release -Duse_mine=OFF -Ddebug=$DEBUG_VERSION -Dprofiling=OFF -Dproduction=ON ..
+cmake -DCMAKE_BUILD_TYPE=Release -Duse_mine=OFF -Denable_extensions=$EXTENSIONS -Ddebug=$DEBUG_VERSION -Dprofiling=OFF -Dproduction=ON ..
 make -j4
 
 echo "Creating $PACK.tar.gz ..."
@@ -59,25 +66,27 @@ rm -rf $PACK
 $SHASUM $PACK.tar.gz
 echo `pwd`/$PACK.tar.gz
 
-echo "Creating $PACK_EXT.tar.gz ..."
-mkdir $PACK_EXT
-mkdir $PACK_EXT/include
-mkdir $PACK_EXT/include/residue/
-mkdir $PACK_EXT/include/residue/extensions/
-mkdir $PACK_EXT/include/residue/core/
-mkdir $PACK_EXT/include/residue/gason/
-cp libresidue-extension-st.a $PACK_EXT/libresidue-extension-st.$VERSION.a
-cp libresidue-extension.$VERSION.* $PACK_EXT/
-cp ../include/extensions.h $PACK_EXT/include/residue/
-cp ../src/extensions/extension.h $PACK_EXT/include/residue/extensions/
-cp ../src/extensions/log-extension.h $PACK_EXT/include/residue/extensions/
-cp ../src/extensions/pre-archive-extension.h $PACK_EXT/include/residue/extensions/
-cp ../src/extensions/post-archive-extension.h $PACK_EXT/include/residue/extensions/
-cp ../src/core/json-doc.h $PACK_EXT/include/residue/core/
-cp ../deps/gason/gason.h $PACK_EXT/include/residue/gason/
+if [ "$EXTENSIONS" = "ON" ];then
+    echo "Creating $PACK_EXT.tar.gz ..."
+    mkdir $PACK_EXT
+    mkdir $PACK_EXT/include
+    mkdir $PACK_EXT/include/residue/
+    mkdir $PACK_EXT/include/residue/extensions/
+    mkdir $PACK_EXT/include/residue/core/
+    mkdir $PACK_EXT/include/residue/gason/
+    cp libresidue-extension-st.a $PACK_EXT/libresidue-extension-st.$VERSION.a
+    cp libresidue-extension.$VERSION.* $PACK_EXT/
+    cp ../include/extensions.h $PACK_EXT/include/residue/
+    cp ../src/extensions/extension.h $PACK_EXT/include/residue/extensions/
+    cp ../src/extensions/log-extension.h $PACK_EXT/include/residue/extensions/
+    cp ../src/extensions/pre-archive-extension.h $PACK_EXT/include/residue/extensions/
+    cp ../src/extensions/post-archive-extension.h $PACK_EXT/include/residue/extensions/
+    cp ../src/core/json-doc.h $PACK_EXT/include/residue/core/
+    cp ../deps/gason/gason.h $PACK_EXT/include/residue/gason/
 
-ls -lh $PACK_EXT
-tar cfz $PACK_EXT.tar.gz $PACK_EXT
-rm -rf $PACK_EXT
-$SHASUM $PACK_EXT.tar.gz
-echo `pwd`/$PACK_EXT.tar.gz
+    ls -lh $PACK_EXT
+    tar cfz $PACK_EXT.tar.gz $PACK_EXT
+    rm -rf $PACK_EXT
+    $SHASUM $PACK_EXT.tar.gz
+    echo `pwd`/$PACK_EXT.tar.gz
+fi
